Remove text-based UI remnants
